Edited RSS Feed/Blog link to merge in two separate blogs - my official one I use for Latinforimagination and a backup one I can post class assignments to. This should give me more freedom to do assignments and to more concise posts about what I'm doing/have been experimenting with recently, without spending a lot of time worrying about making them a perfect fit for the official blog.
Overlap should be less of an issue to deal with, and more of an opportunity with new posts.
The hope is that for some topics, ie, the research I'm doing on FOSS design practices, I can write less formal posts on the backup blog and link to things like transcripts - that will give people in the class/my network more of an opportunity to weigh in on what's going on.
When I'm at the point where I want to write a formal paper to post on the blog/gamasutra, I can do that without worrying about readers already having been flooded with the content.

Signed-off-by: Daniel Shumway <shumway.danny@gmail.com>